Endpoint paths readjusted and adviceApi endpoints moved to another controller.
